APY-142 : Rewrite packiging code
Repaired the setups init for packaging, added also basic
rest distribution.

+++ distribution-manager/ally-distribution/ally/distribution/packaging/impl/processor/build.py
@@ -9,15 +9,15 @@ Created on Feb 10, 2014
 Builds a package to an egg.
 '''
 
+from distutils.core import run_setup
 import logging
+import os
+from shutil import rmtree
 
 from ally.container.ioc import injected
 from ally.design.processor.attribute import requires
 from ally.design.processor.context import Context
 from ally.design.processor.handler import HandlerProcessor
-import os
-from shutil import rmtree
-from distutils.core import run_setup
 
 
 # --------------------------------------------------------------------
@@ -25,6 +25,13 @@ log = logging.getLogger(__name__)
 
 # --------------------------------------------------------------------
 
+class Distribution(Context):
+    '''
+    The distribution context.
+    '''
+    # ---------------------------------------------------------------- Required
+    packages = requires(list)
+    
 class Package(Context):
     '''
     The package context.
@@ -47,26 +54,31 @@ class BuildHandler(HandlerProcessor):
     
     def __init__(self):
         assert isinstance(self.pathBuild, str), 'Invalid build path %s' % self.pathBuild
-        super().__init__()
+        super().__init__(Package=Package)
 
-    def process(self, chain, package:Package, **keyargs):
+    def process(self, chain, distribution:Distribution, **keyargs):
         '''
         @see: HandlerProcessor.process
         
         Provides the package build.
         '''
-        assert isinstance(package, Package), 'Invalid package %s' % package
-        
+        assert isinstance(distribution, Distribution), 'Invalid distribution %s' % distribution
+        if not distribution.packages: return
+            
         savedCwd = os.getcwd()
-        os.chdir(package.path)
-        
-        buildPath = os.path.join(savedCwd, self.pathBuild)
-        run_setup(package.pathSetupPy, ('sdist', '--dist-dir', buildPath, 'bdist_egg', '--dist-dir', buildPath))
-        # run_setup(package.pathSetupPy, ('--help', 'sdist'))
-        
-        # Cleaning setup directories.
-        rmtree('build')
-        rmtree('%s.egg-info' % package.name)
-        
+             
+        for package in distribution.packages:
+            assert isinstance(package, Package), 'Invalid package %s' % package
+
+            os.chdir(package.path)
+            
+            buildPath = os.path.join(savedCwd, self.pathBuild)
+            run_setup(package.pathSetupPy, ('sdist', '--dist-dir', buildPath, 'bdist_egg', '--dist-dir', buildPath))
+            # run_setup(package.pathSetupPy, ('--help', 'sdist'))
+            
+            # Cleaning setup directories.
+            rmtree('build')
+            rmtree('%s.egg-info' % package.name.replace('-', '_'))
+            
         # Restoring environment. 
         os.chdir(savedCwd)
